pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "contractions>=0.1.73,<1.0",
    "datasets>=2.14,<3.0",
    "fastapi>=0.104,<1.0",
    "gunicorn>=21.0,<23.0",
    "huggingface-hub>=0.20,<1.0",
    "joblib>=1.3,<2.0",
    "langdetect>=1.0.9,<2.0",
    "lime>=0.2,<1.0",
    "matplotlib>=3.7,<4.0",
    "networkx>=3.0,<4.0",
    "nltk>=3.8,<4.0",
    "numpy>=1.24,<2.0",
    "pandas>=2.0,<3.0",
    "portalocker>=2.0,<4.0",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "reportlab>=4.0,<5.0",
    "scikit-learn>=1.3,<2.0",
    "scipy>=1.10,<2.0",
    "seaborn>=0.12,<1.0",
    "sentence-transformers>=2.2,<4.0",
    "shap>=0.43,<1.0",
    "spacy>=3.7,<4.0",
    "tarsafe>=0.0.5",
    "torch>=2.1,<3.0",
    "tqdm>=4.66,<5.0",
    "transformers>=4.35,<5.0",
    "uvicorn[standard]>=0.24,<1.0",
    "vadersentiment>=3.3,<4.0",
    "wordcloud>=1.9,<2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
